+++ adminer/create.inc.php
@@ -168,6 +168,13 @@ foreach ($engines as $engine) {
 		break;
 	}
 }
+
+$max_columns = max_input_vars(12, 20); // 12 - the maximum number of inputs per column, 20 - the other inputs
+if ($max_columns) {
+	// the message is printed also if the number of columns is fine, JavaScript displays it after adding more columns
+	$hidden = (count($row["fields"]) > $max_columns ? "" : " hidden");
+	echo "<p" . ($hidden ? " id='max-fields' data-columns='$max_columns'" : "") . " class='error$hidden'>" . max_input_vars_error() . "\n";
+}
 ?>
 
 <form action="" method="post" id="form">