+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Alter table: Move columns on touch screens
 - Alter table: Edit multiline column comments in a textarea
 - Alter table, Privileges: Do not print a success message if nothing was changed
+- Alter table, Select: Warn if the number of fields can exceed max_input_vars
 - Select: Remove asking for the page number after clicking the Page legend
 - Select: Highlight the current page in the pagination
 - Select: Fix the number of selected rows after going back in history